Rename msd to displacement, fix imports.
