+++ tests/test_trajectory_displacement.py
